this commit changes how random poly of weight 0 is generated. The algorithm is now guaranteed to finnish if finite time. It needs to be tested with proper q and k settings though. I am almost certain that the current implementation WILL NOT produce uniform distribution, which may interfere with DFR.
this commit also changes the way encoding w/o matrix is done. In the previous version, g transpose poly was calculated and it was shifted right mod x^k-1 k times. The new implementation does not calculate g transpose at all. Instead, the coefficients of g are stored in the vector g_coeffs. During the encoding, this vector is iterated in reverse starting from a fixed position. When the start of the vector is reached, the iteration continues from the end of the vector. This way of encoding should be much more memory efficient than the original g_transpose one, but also much faster.